Added features on nsm-cli and example - Added option to pass attestion document parameter as plaintext - Added `GetRandom` function to nsm-cli - Generate RSA keypair inside enclave and attach into attestation - Change docker base to `python:slim`

+++ example/server/build.sh
@@ -2,7 +2,10 @@
 
 set -e
 
+if [[ -f nsm-cli-demo.eif ]]; then
 rm nsm-cli-demo.eif
+fi
+
 docker rmi -f nsm-cli-demo:latest
 docker build ../.. -f Dockerfile --build-arg ARCH=$(uname -m) -t nsm-cli-demo:latest
